// ui-service/src/app/components/mapping-case/model/mapping-case-column-model.ts
export enum MappingCaseColumn {
  ID = 'id',
  HOME_TEAM_NAME = 'homeTeamName',
  AWAY_TEAM_NAME = 'awayTeamName',
  MATCH_TEAMS = 'matchTeams',
  AWAY_SIMILARITY_FACTOR = 'awaySimilarityFactor',
  HOME_SIMILARITY_FACTOR = 'homeSimilarityFactor',
  SOURCE_SYSTEM_NAME = 'sourceSystemName',
  STATUS = 'status',
  CREATED_ON = 'createdOn',
  UPDATED_ON = 'updatedOn',
  CREATED_BY = 'createdBy',
  UPDATED_BY = 'updatedBy',
  ACCEPT = 'accept',
  REJECT = 'reject'
}